Support to prune unreachable nodes
in case of pssh.exceptions.ConnectionError exception and
the node is unreachable. Unreachability is ensured
by creating a ssh session to the specific set of nodes
which raised ConnectionError.

in case of pssh.exceptions.Timeout exception and
the node is unreachable. Unreachability is ensured
by creating a ssh session to the specific set of nodes
which raised Timeout.

Only ConnectionError and AuthError exceptions are masked by
stop_on_errors natively in pssh. So to mask Timeout exception
during stop_on_error, custom implementation is done in cvs
parallel_ssh_lib.py, where we catch Timeout id stop_on_errors is
set to False and update outout generator with this exception else
just re-raise it.

Added UT to cover these cases

Signed-off-by: Ignatious Johnson <ichristo@amd.com>